Move file-related helpers to own file, set badbit exceptions on ofstreams
